+++ cmd/config/chain_id.go
@@ -1,0 +1,31 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package config
+
+import (
+	"path/filepath"
+
+	"github.com/cosmos/cosmos-sdk/client/config"
+	"github.com/spf13/viper"
+)
+
+// GetChainIDFromHome returns the chain ID from the client configuration
+// in the given home directory.
+func GetChainIDFromHome(home string) (string, error) {
+	v := viper.New()
+	v.AddConfigPath(filepath.Join(home, "config"))
+	v.SetConfigName("client")
+	v.SetConfigType("toml")
+
+	if err := v.ReadInConfig(); err != nil {
+		return "", err
+	}
+	conf := new(config.ClientConfig)
+
+	if err := v.Unmarshal(conf); err != nil {
+		return "", err
+	}
+
+	return conf.ChainID, nil
+}