+++ example_chain/upgrades.go
@@ -1,0 +1,8 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package example_chain
+
+func (app ExampleChain) RegisterUpgradeHandlers() {
+	// No upgrades registered yet
+}